Moved a few more utilities functions around
