+++ NuRadioReco/utilities/trace_utilities.py
@@ -1,5 +1,4 @@
-from NuRadioReco.utilities import units, fft, signal_processing
-import NuRadioReco.framework.base_trace
+from NuRadioReco.utilities import units, signal_processing
 
 import numpy as np
 import scipy.stats
@@ -24,6 +23,26 @@ def get_efield_antenna_factor(*args, **kwargs):
 def get_channel_voltage_from_efield(*args, **kwargs):
     warnings.warn("get_channel_voltage_from_efield is moved to NuRadioReco.utilities.signal_processing.get_channel_voltage_from_efield", DeprecationWarning)
     return signal_processing.get_channel_voltage_from_efield(*args, **kwargs)
+
+def upsampling_fir(*args, **kwargs):
+    warnings.warn("upsampling_fir is moved to NuRadioReco.utilities.signal_processing.upsampling_fir", DeprecationWarning)
+    return signal_processing.upsampling_fir(*args, **kwargs)
+
+
+def butterworth_filter_trace(*args, **kwargs):
+    warnings.warn("butterworth_filter_trace is moved to NuRadioReco.utilities.signal_processing.butterworth_filter_trace", DeprecationWarning)
+    return signal_processing.butterworth_filter_trace(*args, **kwargs)
+
+
+def apply_butterworth(*args, **kwargs):
+    warnings.warn("apply_butterworth is moved to NuRadioReco.utilities.signal_processing.apply_butterworth", DeprecationWarning)
+    return signal_processing.apply_butterworth(*args, **kwargs)
+
+
+def delay_trace(*args, **kwargs):
+    warnings.warn("delay_trace is moved to NuRadioReco.utilities.signal_processing.delay_trace", DeprecationWarning)
+    return signal_processing.delay_trace(*args, **kwargs)
+
 
 
 def get_electric_field_energy_fluence(electric_field_trace, times, signal_window_mask=None, noise_window_mask=None):
@@ -177,208 +196,6 @@ def get_stokes(trace_u, trace_v, window_samples=128, squeeze=True):
     if squeeze:
         return np.squeeze(stokes)
     return stokes
-
-
-def upsampling_fir(trace, original_sampling_frequency, int_factor=2, ntaps=2 ** 7):
-    """
-    This function performs an upsampling by inserting a number of zeroes
-    between samples and then applying a finite impulse response (FIR) filter.
-
-    Parameters
-    ----------
-
-    trace: array of floats
-        Trace to be upsampled
-    original_sampling_frequency: float
-        Sampling frequency of the input trace
-    int_factor: integer
-        Upsampling factor. The resulting trace will have a sampling frequency
-        int_factor times higher than the original one
-    ntaps: integer
-        Number of taps (order) of the FIR filter
-
-    Returns
-    -------
-    upsampled_trace: array of floats
-        The upsampled trace
-    """
-
-    if (np.abs(int(int_factor) - int_factor) > 1e-3):
-        warning_msg = "The input upsampling factor does not seem to be close to an integer."
-        warning_msg += "It has been rounded to {}".format(int(int_factor))
-        logger.warning(warning_msg)
-
-    int_factor = int(int_factor)
-
-    if (int_factor <= 1):
-        error_msg = "Upsampling factor is less or equal to 1. Upsampling will not be performed."
-        raise ValueError(error_msg)
-
-    zeroed_trace = np.zeros(len(trace) * int_factor)
-    for i_point, point in enumerate(trace[:-1]):
-        zeroed_trace[i_point * int_factor] = point
-
-    upsampled_delta_time = 1 / (int_factor * original_sampling_frequency)
-    upsampled_times = np.arange(0, len(zeroed_trace) * upsampled_delta_time, upsampled_delta_time)
-
-    cutoff = 1. / int_factor
-    fir_coeffs = scipy.signal.firwin(ntaps, cutoff, window='boxcar')
-    upsampled_trace = np.convolve(zeroed_trace, fir_coeffs)[:len(upsampled_times)] * int_factor
-
-    return upsampled_trace
-
-
-def butterworth_filter_trace(trace, sampling_frequency, passband, order=8):
-    """
-    Filters a trace using a Butterworth filter.
-
-    Parameters
-    ----------
-
-    trace: array of floats
-        Trace to be filtered
-    sampling_frequency: float
-        Sampling frequency
-    passband: (float, float) tuple
-        Tuple indicating the cutoff frequencies
-    order: integer
-        Filter order
-
-    Returns
-    -------
-
-    filtered_trace: array of floats
-        The filtered trace
-    """
-
-    n_samples = len(trace)
-
-    spectrum = fft.time2freq(trace, sampling_frequency)
-    frequencies = fft.freqs(n_samples, sampling_frequency)
-
-    filtered_spectrum = apply_butterworth(spectrum, frequencies, passband, order)
-    filtered_trace = fft.freq2time(filtered_spectrum, sampling_frequency)
-
-    return filtered_trace
-
-
-def apply_butterworth(spectrum, frequencies, passband, order=8):
-    """
-    Calculates the response from a Butterworth filter and applies it to the
-    input spectrum
-
-    Parameters
-    ----------
-    spectrum: array of complex
-        Fourier spectrum to be filtere
-    frequencies: array of floats
-        Frequencies of the input spectrum
-    passband: (float, float) tuple
-        Tuple indicating the cutoff frequencies
-    order: integer
-        Filter order
-
-    Returns
-    -------
-    filtered_spectrum: array of complex
-        The filtered spectrum
-    """
-
-    f = np.zeros_like(frequencies, dtype=complex)
-    mask = frequencies > 0
-    b, a = scipy.signal.butter(order, passband, 'bandpass', analog=True)
-    w, h = scipy.signal.freqs(b, a, frequencies[mask])
-    f[mask] = h
-
-    filtered_spectrum = f * spectrum
-
-    return filtered_spectrum
-
-
-def delay_trace(trace, sampling_frequency, time_delay, crop_trace=True):
-    """
-    Delays a trace by transforming it to frequency and multiplying by phases.
-
-    A positive delay means that the trace is shifted to the right, i.e., its delayed.
-    A negative delay would mean that the trace is shifted to the left. Since this
-    method is cyclic, the delayed trace will have unphysical samples at either the
-    beginning (delayed, positive `time_delay`) or at the end (negative `time_delay`).
-    Those samples can be cropped (optional, default=True).
-
-    Parameters
-    ----------
-    trace: array of floats or `NuRadioReco.framework.base_trace.BaseTrace`
-        Array containing the trace
-    sampling_frequency: float
-        Sampling rate for the trace
-    time_delay: float
-        Time delay used for transforming the trace. Must be positive or 0
-    crop_trace: bool (default: True)
-        If True, the trace is cropped to remove samples what are unphysical
-        after delaying (rolling) the trace.
-
-    Returns
-    -------
-    delayed_trace: array of floats
-        The delayed, cropped trace
-    dt_start: float (optional)
-        The delta t of the trace start time. Only returned if crop_trace is True.
-    """
-    # Do nothing if time_delay is 0
-    if not time_delay:
-        if isinstance(trace, NuRadioReco.framework.base_trace.BaseTrace):
-            if crop_trace:
-                return trace.get_trace(), 0
-            else:
-                return trace.get_trace()
-        else:
-            if crop_trace:
-                return trace, 0
-            else:
-                return trace
-
-    if isinstance(trace, NuRadioReco.framework.base_trace.BaseTrace):
-        spectrum = trace.get_frequency_spectrum()
-        frequencies = trace.get_frequencies()
-        if trace.get_sampling_rate() != sampling_frequency:
-            raise ValueError("The sampling frequency of the trace does not match the given sampling frequency.")
-    else:
-        n_samples = len(trace)
-        spectrum = fft.time2freq(trace, sampling_frequency)
-        frequencies = fft.freqs(n_samples, sampling_frequency)
-
-    spectrum *= np.exp(-1j * 2 * np.pi * frequencies * time_delay)
-
-    delayed_trace = fft.freq2time(spectrum, sampling_frequency)
-    cycled_samples = int(round(time_delay * sampling_frequency))
-
-    if crop_trace:
-        # according to a NuRadio convention, traces should have an even number of samples.
-        # Make sure that after cropping the trace has an even number of samples (assuming that it was even before).
-        if cycled_samples % 2 != 0:
-            cycled_samples += 1
-
-        if time_delay >= 0:
-            delayed_trace = delayed_trace[cycled_samples:]
-            dt_start = cycled_samples * sampling_frequency
-        else:
-            delayed_trace = delayed_trace[:-cycled_samples]
-            dt_start = 0
-
-        return delayed_trace, dt_start
-
-    else:
-        # Check if unphysical samples contain any signal and if so, throw a warning
-        if time_delay > 0:
-            if np.any(np.abs(delayed_trace[:cycled_samples]) > 0.01 * units.microvolt):
-                logger.warning("The delayed trace has unphysical samples that contain signal. "
-                    "Consider cropping the trace to remove these samples.")
-        else:
-            if np.any(np.abs(delayed_trace[-cycled_samples:]) > 0.01 * units.microvolt):
-                logger.warning("The delayed trace has unphysical samples that contain signal. "
-                    "Consider cropping the trace to remove these samples.")
-
-        return delayed_trace
 
 
 def peak_to_peak_amplitudes(trace, coincidence_window_size):
